pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "embeddingviz"
description = "Visualization of music-text embeddings spaces "
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Philip Tovstogan", email = "phil.tgun@gmail.com" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dynamic = ["version"]

dependencies = [
    "streamlit~=1.24",
    "requests~=2.31",
    "python-dotenv~=1.0",
    "transformers~=4.31",
    "torch~=2.0",
    "plotly~=5.15",
    "pandas~=2.1",
    "streamlit-audio-plot~=0.0",
]

[project.optional-dependencies]
dev = [
    # code quality
    "ruff>=0.1.5",
    "pre-commit>=3.2.0",
    "mypy>=1.1.1",
    # local dev
    "types-requests~=2.31",
    "pandas-stubs~=2.1",
]
test = []


[tool.hatch.version]
path = "embeddingviz/__init__.py"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[[tool.mypy.overrides]]
module = ["transformers", "torchaudio", "networkx", "matplotlib"]
ignore_missing_imports = true
